requirements.txt
fastapi
uvicorn
gspread
google-auth
google-api-python-client
python-multipart
pydantic
matplotlib
numpy
